Rename device-list section to devices
