+++ public/app/screens/main/sections/devices/device-finder/device-finder-dialog/index.js
