+++ public/app/screens/main/sections/devices/device-finder/index.js
