+++ public/app/screens/main/sections/devices/device-finder/style.scss
