+++ public/app/screens/main/sections/devices/device-finder/test/device-finder.spec.jsx
