+++ public/app/screens/main/sections/devices/index.js
