+++ public/app/screens/main/sections/devices/room/device/device.jsx
