+++ public/app/screens/main/sections/devices/room/device/index.js
