+++ public/app/screens/main/sections/devices/room/device/style.scss
