+++ public/app/screens/main/sections/devices/room/device/test/device.spec.jsx
@@ -1,4 +1,4 @@
-var Device = require('screens/main/sections/device-list/room/device/device.jsx');
+var Device = require('screens/main/sections/devices/room/device/device.jsx');
 var React = require('react/addons');
 var TestUtils = React.addons.TestUtils;
