+++ public/app/screens/main/sections/devices/room/gateway/gateway.jsx
