+++ public/app/screens/main/sections/devices/room/gateway/index.js
