+++ public/app/screens/main/sections/devices/room/room.jsx
